TGP-3176: refactor goodsRecord goodsDescriptionController into 2 separate controillers and move folders
