+++ app/controllers/goodsRecord/goodsDescription/HasGoodsDescriptionChangeController.scala
@@ -1,0 +1,82 @@
+/*
+ * Copyright 2024 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package controllers.goodsRecord.goodsDescription
+
+import controllers.BaseController
+import controllers.actions._
+import forms.goodsRecord.HasGoodsDescriptionChangeFormProvider
+import models.Mode
+import models.helper.GoodsDetailsUpdate
+import navigation.GoodsRecordNavigator
+import pages.goodsRecord.HasGoodsDescriptionChangePage
+import play.api.i18n.MessagesApi
+import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
+import repositories.SessionRepository
+import services.AuditService
+import utils.SessionData.{dataRemoved, dataUpdated, pageUpdated}
+import views.html.goodsRecord.HasGoodsDescriptionChangeView
+
+import javax.inject.Inject
+import scala.annotation.unused
+import scala.concurrent.{ExecutionContext, Future}
+
+class HasGoodsDescriptionChangeController @Inject() (
+  override val messagesApi: MessagesApi,
+  sessionRepository: SessionRepository,
+  navigator: GoodsRecordNavigator,
+  identify: IdentifierAction,
+  getData: DataRetrievalAction,
+  requireData: DataRequiredAction,
+  profileAuth: ProfileAuthenticateAction,
+  auditService: AuditService,
+  formProvider: HasGoodsDescriptionChangeFormProvider,
+  val controllerComponents: MessagesControllerComponents,
+  view: HasGoodsDescriptionChangeView
+)(implicit @unused ec: ExecutionContext)
+    extends BaseController {
+
+  private val form = formProvider()
+
+  def onPageLoad(mode: Mode, recordId: String): Action[AnyContent] =
+    (identify andThen profileAuth andThen getData andThen requireData) { implicit request =>
+      auditService
+        .auditStartUpdateGoodsRecord(
+          request.eori,
+          request.affinityGroup,
+          GoodsDetailsUpdate,
+          recordId
+        )
+
+      val preparedForm = prepareForm(HasGoodsDescriptionChangePage(recordId), form)
+
+      Ok(view(preparedForm, mode, recordId)).removingFromSession(dataRemoved, dataUpdated, pageUpdated)
+    }
+
+  def onSubmit(mode: Mode, recordId: String): Action[AnyContent] =
+    (identify andThen profileAuth andThen getData andThen requireData).async { implicit request =>
+      form
+        .bindFromRequest()
+        .fold(
+          formWithErrors => Future.successful(BadRequest(view(formWithErrors, mode, recordId))),
+          value =>
+            for {
+              updatedAnswers <- Future.fromTry(request.userAnswers.set(HasGoodsDescriptionChangePage(recordId), value))
+              _              <- sessionRepository.set(updatedAnswers)
+            } yield Redirect(navigator.nextPage(HasGoodsDescriptionChangePage(recordId), mode, updatedAnswers))
+        )
+    }
+}
