+++ app/views/goodsRecord/HasGoodsDescriptionChangeView.scala.html
@@ -36,7 +36,7 @@
 
     <p class="govuk-body">@messages("hasGoodsDescriptionChange.p1")</p>
 
-    @formHelper(action = controllers.goodsRecord.routes.HasGoodsDescriptionChangeController.onSubmit(mode, recordId), Symbol("autoComplete") -> "off") {
+    @formHelper(action = controllers.goodsRecord.goodsDescription.routes.HasGoodsDescriptionChangeController.onSubmit(mode, recordId), Symbol("autoComplete") -> "off") {
 
         @govukRadios(
             RadiosViewModel.yesNo(
