+++ conf/app.goodsRecord.routes
@@ -24,15 +24,15 @@ POST        /create-record/start                                                
 
 GET         /create-record/:recordId/success                                                        controllers.goodsRecord.CreateRecordSuccessController.onPageLoad(recordId: String)
 
-GET         /create-record/goods-description                                                        controllers.goodsRecord.GoodsDescriptionController.onPageLoadCreate(mode: Mode = NormalMode)
-POST        /create-record/goods-description                                                        controllers.goodsRecord.GoodsDescriptionController.onSubmitCreate(mode: Mode = NormalMode)
-GET         /create-record/goods-description/check                                                  controllers.goodsRecord.GoodsDescriptionController.onPageLoadCreate(mode: Mode = CheckMode)
-POST        /create-record/goods-description/check                                                  controllers.goodsRecord.GoodsDescriptionController.onSubmitCreate(mode: Mode = CheckMode)
+GET         /create-record/goods-description                                                        controllers.goodsRecord.goodsDescription.CreateGoodsDescriptionController.onPageLoad(mode: Mode = NormalMode)
+POST        /create-record/goods-description                                                        controllers.goodsRecord.goodsDescription.CreateGoodsDescriptionController.onSubmit(mode: Mode = NormalMode)
+GET         /create-record/goods-description/check                                                  controllers.goodsRecord.goodsDescription.CreateGoodsDescriptionController.onPageLoad(mode: Mode = CheckMode)
+POST        /create-record/goods-description/check                                                  controllers.goodsRecord.goodsDescription.CreateGoodsDescriptionController.onSubmit(mode: Mode = CheckMode)
 
-GET         /update-record/:recordId/goods-description                                              controllers.goodsRecord.GoodsDescriptionController.onPageLoadUpdate(mode: Mode = NormalMode, recordId: String)
-POST        /update-record/:recordId/goods-description                                              controllers.goodsRecord.GoodsDescriptionController.onSubmitUpdate(mode: Mode = NormalMode, recordId: String)
-GET         /update-record/:recordId/goods-description/check                                        controllers.goodsRecord.GoodsDescriptionController.onPageLoadUpdate(mode: Mode = CheckMode, recordId: String)
-POST        /update-record/:recordId/goods-description/check                                        controllers.goodsRecord.GoodsDescriptionController.onSubmitUpdate(mode: Mode = CheckMode, recordId: String)
+GET         /update-record/:recordId/goods-description                                              controllers.goodsRecord.goodsDescription.UpdateGoodsDescriptionController.onPageLoad(mode: Mode = NormalMode, recordId: String)
+POST        /update-record/:recordId/goods-description                                              controllers.goodsRecord.goodsDescription.UpdateGoodsDescriptionController.onSubmit(mode: Mode = NormalMode, recordId: String)
+GET         /update-record/:recordId/goods-description/check                                        controllers.goodsRecord.goodsDescription.UpdateGoodsDescriptionController.onPageLoad(mode: Mode = CheckMode, recordId: String)
+POST        /update-record/:recordId/goods-description/check                                        controllers.goodsRecord.goodsDescription.UpdateGoodsDescriptionController.onSubmit(mode: Mode = CheckMode, recordId: String)
 
 GET         /create-record/check-your-answers                                                       controllers.goodsRecord.CyaCreateRecordController.onPageLoad()
 POST        /create-record/check-your-answers                                                       controllers.goodsRecord.CyaCreateRecordController.onSubmit()
@@ -59,10 +59,10 @@ POST        /update-record/:recordId/change-record-warning/country-of-origin    
 GET         /update-record/:recordId/change-record-warning/country-of-origin/check                  controllers.goodsRecord.countryOfOrigin.HasCountryOfOriginChangeController.onPageLoad(mode: Mode = CheckMode, recordId: String)
 POST        /update-record/:recordId/change-record-warning/country-of-origin/check                  controllers.goodsRecord.countryOfOrigin.HasCountryOfOriginChangeController.onSubmit(mode: Mode = CheckMode, recordId: String)
 
-GET         /update-record/:recordId/change-record-warning/goods-description                        controllers.goodsRecord.HasGoodsDescriptionChangeController.onPageLoad(mode: Mode = NormalMode, recordId: String)
-POST        /update-record/:recordId/change-record-warning/goods-description                        controllers.goodsRecord.HasGoodsDescriptionChangeController.onSubmit(mode: Mode = NormalMode, recordId: String)
-GET         /update-record/:recordId/change-record-warning/goods-description/check                  controllers.goodsRecord.HasGoodsDescriptionChangeController.onPageLoad(mode: Mode = CheckMode, recordId: String)
-POST        /update-record/:recordId/change-record-warning/goods-description/check                  controllers.goodsRecord.HasGoodsDescriptionChangeController.onSubmit(mode: Mode = CheckMode, recordId: String)
+GET         /update-record/:recordId/change-record-warning/goods-description                        controllers.goodsRecord.goodsDescription.HasGoodsDescriptionChangeController.onPageLoad(mode: Mode = NormalMode, recordId: String)
+POST        /update-record/:recordId/change-record-warning/goods-description                        controllers.goodsRecord.goodsDescription.HasGoodsDescriptionChangeController.onSubmit(mode: Mode = NormalMode, recordId: String)
+GET         /update-record/:recordId/change-record-warning/goods-description/check                  controllers.goodsRecord.goodsDescription.HasGoodsDescriptionChangeController.onPageLoad(mode: Mode = CheckMode, recordId: String)
+POST        /update-record/:recordId/change-record-warning/goods-description/check                  controllers.goodsRecord.goodsDescription.HasGoodsDescriptionChangeController.onSubmit(mode: Mode = CheckMode, recordId: String)
 
 GET         /update-record/:recordId/check-your-answers/country-of-origin                           controllers.goodsRecord.CyaUpdateRecordController.onPageLoadCountryOfOrigin(recordId: String)
 POST        /update-record/:recordId/check-your-answers/country-of-origin                           controllers.goodsRecord.CyaUpdateRecordController.onSubmitCountryOfOrigin(recordId: String)
