+++ test/viewmodels/checkAnswers/GoodsDescriptionSummarySpec.scala
@@ -46,8 +46,8 @@ class GoodsDescriptionSummarySpec extends SpecBase {
           GoodsDescriptionSummary.rowUpdate(recordForTestingSummaryRows, testRecordId, NormalMode, recordLocked = false)
 
         row.actions mustBe defined
-        row.actions.value.items.head.href mustEqual controllers.goodsRecord.routes.GoodsDescriptionController
-          .onPageLoadUpdate(NormalMode, testRecordId)
+        row.actions.value.items.head.href mustEqual controllers.goodsRecord.goodsDescription.routes.UpdateGoodsDescriptionController
+          .onPageLoad(NormalMode, testRecordId)
           .url
       }
 
@@ -59,7 +59,7 @@ class GoodsDescriptionSummarySpec extends SpecBase {
           GoodsDescriptionSummary.rowUpdate(recordWithAdviceProvided, testRecordId, NormalMode, recordLocked = false)
 
         row.actions mustBe defined
-        row.actions.value.items.head.href mustEqual controllers.goodsRecord.routes.HasGoodsDescriptionChangeController
+        row.actions.value.items.head.href mustEqual controllers.goodsRecord.goodsDescription.routes.HasGoodsDescriptionChangeController
           .onPageLoad(NormalMode, testRecordId)
           .url
       }
